run: package naming and minor test fixes
