+++ run/markdown-preview/renderer/go.mod
@@ -3,9 +3,9 @@ module github.com/GoogleCloudPlatform/golang-samples/run/markdown-preview/render
 go 1.13
 
 require (
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200708192522-c9d21016c9ac
 	github.com/microcosm-cc/bluemonday v1.0.3
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/russross/blackfriday/v2 v2.0.1
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
-	google.golang.org/api v0.28.0
+	golang.org/x/net v0.0.0-20200625001655-4c5254603344 // indirect
 )
